Continued Refactoring of GUI Code - WIP
